refactor: components -> shared로 폴더명 변경
