+++ frontend/src/features/post/components/CreatePostHeader.tsx
@@ -1,4 +1,4 @@
-import Button from "@/components/ui/Button";
+import Button from "@/shared/ui/Button";
 
 export default function CreatePostHeader() {
   return (
