+++ frontend/src/features/search/api/getPullRequests.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
 
+import type { PullRequest } from "@/entities/pullrequest";
 import type { QueryConfig } from "@/services/react-query";
 import { getRequest } from "@/services/request";
-import type { PullRequest } from "@/types/pullrequest";
 
 const BASE_URL = "/github";
