+++ frontend/src/shared/ui/Badge.tsx
