+++ frontend/src/shared/ui/Divider.tsx
