+++ frontend/src/shared/ui/Fallback.tsx
